rematch2-package {rematch2}	R Documentation

Match Regular Expressions with a Nicer 'API'

Description

A small wrapper on 'regexpr' to extract the matches and captured groups from the match of a regular expression to a character vector. See re_match.

Author(s)

Maintainer: Gábor Csárdi csardi.gabor@gmail.com

Other contributors:

Matthew Lincoln matthew.d.lincoln@gmail.com [contributor]

See Also

Useful links:

https://github.com/r-lib/rematch2#readme

Report bugs at https://github.com/r-lib/rematch2/issues

[Package rematch2 version 2.1.2 Index]